+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/Instrumentable.kt
@@ -3,17 +3,18 @@
 package io.sentry.android.gradle.instrumentation
 
 import com.android.build.api.instrumentation.ClassContext
+import java.io.Serializable
 import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.MethodVisitor
 
-interface Instrumentable<Visitor, InstrumentableContext> {
+interface Instrumentable<Visitor, InstrumentableContext> : Serializable {
 
     /**
      * Fully-qualified name of the instrumentable. Examples:
      * Class: androidx.sqlite.db.framework.FrameworkSQLiteDatabase
      * Method: query
      */
-    val fqName: String
+    val fqName: String get() = ""
 
     /**
      * Provides a visitor for this instrumentable. A visitor can be one of the visitors defined
@@ -30,12 +31,6 @@ interface Instrumentable<Visitor, InstrumentableContext> {
         originalVisitor: Visitor,
         parameters: SpanAddingClassVisitorFactory.SpanAddingParameters
     ): Visitor
-
-    /**
-     * Provides children instrumentables that are going to be used when visiting the current
-     * class/method/field/etc.
-     */
-    val children: List<Instrumentable<*, *>> get() = emptyList()
 
     /**
      * Defines whether this object is instrumentable or not based on [data]