+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/androidx/sqlite/database/visitor/ExecSqlMethodVisitor.kt
@@ -2,6 +2,7 @@ package io.sentry.android.gradle.instrumentation.androidx.sqlite.database.visito
 
 import io.sentry.android.gradle.instrumentation.AbstractSpanAddingMethodVisitor
 import io.sentry.android.gradle.instrumentation.ReturnType
+import io.sentry.android.gradle.instrumentation.SpanOperations
 import io.sentry.android.gradle.instrumentation.util.Types
 import org.objectweb.asm.Label
 import org.objectweb.asm.MethodVisitor
@@ -27,6 +28,7 @@ class ExecSqlMethodVisitor(
         originalVisitor.visitTryCatchBlocks(expectedException = "android/database/SQLException")
 
         originalVisitor.visitStartSpan(gotoIfNull = label0) {
+            visitLdcInsn(SpanOperations.DB_SQL_QUERY)
             visitVarInsn(ALOAD, 1)
         }