+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/BaseSentryPluginTest.kt
@@ -1,0 +1,111 @@
+package io.sentry.android.gradle
+
+import java.io.File
+import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.internal.PluginUnderTestMetadataReading
+import org.junit.Before
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+import org.junit.runners.Parameterized
+
+@Suppress("FunctionName")
+abstract class BaseSentryPluginTest(
+    private val androidGradlePluginVersion: String,
+    private val gradleVersion: String
+) {
+    @get:Rule
+    val testProjectDir = TemporaryFolder()
+
+    private val projectTemplateFolder = File("src/test/resources/testFixtures/appTestProject")
+
+    private lateinit var rootBuildFile: File
+    protected lateinit var appBuildFile: File
+    protected lateinit var runner: GradleRunner
+
+    abstract val additionalRootProjectConfig: String
+
+    @Before
+    fun setup() {
+        projectTemplateFolder.copyRecursively(testProjectDir.root)
+
+        val pluginClasspath = PluginUnderTestMetadataReading.readImplementationClasspath()
+            .joinToString(separator = ", ") { "\"$it\"" }
+            .replace(File.separator, "/")
+
+        appBuildFile = File(testProjectDir.root, "app/build.gradle")
+        rootBuildFile = testProjectDir.writeFile("build.gradle") {
+            // language=Groovy
+            """
+            buildscript {
+              repositories {
+                google()
+                gradlePluginPortal()
+              }
+              dependencies {
+                classpath 'com.android.tools.build:gradle:$androidGradlePluginVersion'
+                // This is needed to populate the plugin classpath instead of using
+                // withPluginClasspath on the Gradle Runner.
+                classpath files($pluginClasspath)
+              }
+            }
+
+            allprojects {
+              repositories {
+                google()
+                mavenCentral()
+              }
+            }
+            subprojects {
+              pluginManager.withPlugin('com.android.application') {
+                android {
+                  compileSdkVersion 30
+                  defaultConfig {
+                    applicationId "com.example"
+                    minSdkVersion 21
+                  }
+                  buildTypes {
+                    release {
+                      minifyEnabled true
+                      proguardFiles("src/release/proguard-rules.pro")
+                    }
+                  }
+                  $additionalRootProjectConfig
+                }
+              }
+            }
+            """.trimIndent()
+        }
+
+        runner = GradleRunner.create()
+            .withProjectDir(testProjectDir.root)
+            .withArguments("--stacktrace")
+            .withPluginClasspath()
+            .withGradleVersion(gradleVersion)
+    }
+
+    companion object {
+
+        @Parameterized.Parameters(name = "AGP {0}, Gradle {1}")
+        @JvmStatic
+        fun parameters() = listOf(
+            // The supported Gradle version can be found here:
+            // https://developer.android.com/studio/releases/gradle-plugin#updating-gradle
+            // The pair is [AGP Version, Gradle Version]
+            arrayOf("7.0.3", "7.0.2"),
+            arrayOf("7.0.3", "7.1.1"),
+            arrayOf("7.0.3", "7.2"),
+            arrayOf("7.1.0-beta02", "7.2"),
+            arrayOf("7.2.0-alpha03", "7.2")
+        )
+
+        internal fun GradleRunner.appendArguments(vararg arguments: String) =
+            withArguments(this.arguments + arguments)
+
+        private fun TemporaryFolder.writeFile(fileName: String, text: () -> String): File {
+            val file = File(root, fileName)
+            file.parentFile.mkdirs()
+            file.writeText(text())
+            return file
+        }
+    }
+}