+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginTest.kt
@@ -1,91 +1,21 @@
 package io.sentry.android.gradle
 
 import java.io.File
+import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
-import org.gradle.testkit.runner.GradleRunner
-import org.gradle.testkit.runner.internal.PluginUnderTestMetadataReading
 import org.junit.Assert.assertThrows
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.TemporaryFolder
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
 @Suppress("FunctionName")
 @RunWith(Parameterized::class)
 class SentryPluginTest(
-    private val androidGradlePluginVersion: String,
-    private val gradleVersion: String
-) {
-    @get:Rule
-    val testProjectDir = TemporaryFolder()
-
-    private val projectTemplateFolder = File("src/test/resources/testFixtures/appTestProject")
-
-    private lateinit var rootBuildFile: File
-    private lateinit var appBuildFile: File
-    private lateinit var runner: GradleRunner
-
-    @Before
-    fun setup() {
-        projectTemplateFolder.copyRecursively(testProjectDir.root)
-
-        val pluginClasspath = PluginUnderTestMetadataReading.readImplementationClasspath()
-            .joinToString(separator = ", ") { "\"$it\"" }
-            .replace(File.separator, "/")
-
-        appBuildFile = File(testProjectDir.root, "app/build.gradle")
-        rootBuildFile = testProjectDir.writeFile("build.gradle") {
-            // language=Groovy
-            """
-            buildscript {
-              repositories {
-                google()
-                gradlePluginPortal()
-              }
-              dependencies {
-                classpath 'com.android.tools.build:gradle:$androidGradlePluginVersion'
-                // This is needed to populate the plugin classpath instead of using
-                // withPluginClasspath on the Gradle Runner.
-                classpath files($pluginClasspath)
-              }
-            }
-
-            allprojects {
-              repositories {
-                google()
-                mavenCentral()
-              }
-            }
-            subprojects {
-              pluginManager.withPlugin('com.android.application') {
-                android {
-                  compileSdkVersion 30
-                  defaultConfig {
-                    applicationId "com.example"
-                    minSdkVersion 21
-                  }
-                  buildTypes {
-                    release {
-                      minifyEnabled true
-                      proguardFiles("src/release/proguard-rules.pro")
-                    }
-                  }
-                }
-              }
-            }
-            """.trimIndent()
-        }
-
-        runner = GradleRunner.create()
-            .withProjectDir(testProjectDir.root)
-            .withArguments("--stacktrace")
-            .withPluginClasspath()
-            .withGradleVersion(gradleVersion)
-    }
+    androidGradlePluginVersion: String,
+    gradleVersion: String
+) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
 
     @Test
     fun `plugin can be applied`() {
@@ -98,7 +28,7 @@ class SentryPluginTest(
                 }
 
                 sentry {
-                  autoUpload = false
+                  autoUploadProguardMapping = false
                   tracingInstrumentation {
                     enabled = false
                   }
@@ -172,6 +102,31 @@ class SentryPluginTest(
     }
 
     @Test
+    fun `does not include a UUID in the APK if includeProguardMapping is off`() {
+        appBuildFile.writeText(
+            // language=Groovy
+            """
+                plugins {
+                  id "com.android.application"
+                  id "io.sentry.android.gradle"
+                }
+
+                sentry {
+                  includeProguardMapping = false
+                }
+            """.trimIndent()
+        )
+
+        runner
+            .appendArguments(":app:assembleRelease")
+            .build()
+
+        assertThrows(AssertionError::class.java) {
+            verifyProguardUuid(testProjectDir.root)
+        }
+    }
+
+    @Test
     fun `creates uploadSentryNativeSymbols task if uploadNativeSymbols is enabled`() {
         applyUploadNativeSymbols()
 
@@ -237,6 +192,49 @@ class SentryPluginTest(
         assertTrue(":app:transformReleaseClassesWithAsm" in build.output)
     }
 
+    @Test
+    fun `applies only DB instrumentables when only DATABASE feature enabled`() {
+        applyTracingInstrumentation(features = setOf(InstrumentationFeature.DATABASE))
+
+        val build = runner
+            .appendArguments(":app:assembleDebug", "--debug")
+            .build()
+
+        assertTrue {
+            "[sentry] Instrumentables: AndroidXSQLiteDatabase, AndroidXSQLiteStatement," +
+                " AndroidXRoomDao" in build.output
+        }
+    }
+
+    @Test
+    fun `applies only FILE_IO instrumentables when only FILE_IO feature enabled`() {
+        applyTracingInstrumentation(features = setOf(InstrumentationFeature.FILE_IO))
+
+        val build = runner
+            .appendArguments(":app:assembleDebug", "--debug")
+            .build()
+
+        assertTrue {
+            "[sentry] Instrumentables: ChainedInstrumentable" in build.output
+        }
+    }
+
+    @Test
+    fun `applies all instrumentables when all features enabled`() {
+        applyTracingInstrumentation(
+            features = setOf(InstrumentationFeature.DATABASE, InstrumentationFeature.FILE_IO)
+        )
+
+        val build = runner
+            .appendArguments(":app:assembleDebug", "--debug")
+            .build()
+
+        assertTrue {
+            "[sentry] Instrumentables: AndroidXSQLiteDatabase, AndroidXSQLiteStatement," +
+                " AndroidXRoomDao, ChainedInstrumentable" in build.output
+        }
+    }
+
     private fun applyUploadNativeSymbols() {
         appBuildFile.writeText(
             // language=Groovy
@@ -247,7 +245,7 @@ class SentryPluginTest(
                 }
 
                 sentry {
-                  autoUpload = false
+                  autoUploadProguardMapping = false
                   uploadNativeSymbols = true
                   tracingInstrumentation {
                     enabled = false
@@ -266,7 +264,7 @@ class SentryPluginTest(
                   id "io.sentry.android.gradle"
                 }
                 sentry {
-                  autoUpload = true
+                  autoUploadProguardMapping = true
                   ignoredVariants = ["$ignoredVariant"]
                   tracingInstrumentation {
                     enabled = false
@@ -276,7 +274,10 @@ class SentryPluginTest(
         )
     }
 
-    private fun applyTracingInstrumentation(tracingInstrumentation: Boolean = true) {
+    private fun applyTracingInstrumentation(
+        tracingInstrumentation: Boolean = true,
+        features: Set<InstrumentationFeature> = setOf()
+    ) {
         appBuildFile.writeText(
             // language=Groovy
             """
@@ -285,39 +286,25 @@ class SentryPluginTest(
                   id "io.sentry.android.gradle"
                 }
 
+                dependencies {
+                  implementation 'io.sentry:sentry-android:5.5.0'
+                }
+
                 sentry {
-                  autoUpload = false
+                  autoUploadProguardMapping = false
                   tracingInstrumentation {
                     enabled = $tracingInstrumentation
+                    features = ${
+            features.joinToString(
+                prefix = "[",
+                postfix = "]"
+            ) { "${it::class.java.canonicalName}.${it.name}" }
+            }
                   }
                 }
             """.trimIndent()
         )
     }
 
-    companion object {
-
-        @Parameterized.Parameters(name = "AGP {0}, Gradle {1}")
-        @JvmStatic
-        fun parameters() = listOf(
-            // The supported Gradle version can be found here:
-            // https://developer.android.com/studio/releases/gradle-plugin#updating-gradle
-            // The pair is [AGP Version, Gradle Version]
-            arrayOf("7.0.3", "7.0.2"),
-            arrayOf("7.0.3", "7.1.1"),
-            arrayOf("7.0.3", "7.2"),
-            arrayOf("7.1.0-beta02", "7.2"),
-            arrayOf("7.2.0-alpha03", "7.3-rc-5")
-        )
-
-        private fun GradleRunner.appendArguments(vararg arguments: String) =
-            withArguments(this.arguments + arguments)
-
-        private fun TemporaryFolder.writeFile(fileName: String, text: () -> String): File {
-            val file = File(root, fileName)
-            file.parentFile.mkdirs()
-            file.writeText(text())
-            return file
-        }
-    }
+    override val additionalRootProjectConfig: String = ""
 }