+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/instrumentation/classloader/GeneratingMissingClassesClassLoader.kt
@@ -1,8 +1,10 @@
 package io.sentry.android.gradle.instrumentation.classloader
 
 import io.sentry.android.gradle.instrumentation.classloader.mapping.deletionDaoMissingClasses
+import io.sentry.android.gradle.instrumentation.classloader.mapping.gmsMapping
 import io.sentry.android.gradle.instrumentation.classloader.mapping.insertionDaoMissingClasses
 import io.sentry.android.gradle.instrumentation.classloader.mapping.selectDaoMissingClasses
+import io.sentry.android.gradle.instrumentation.classloader.mapping.sqliteCopyOpenHelperMissingClasses
 import io.sentry.android.gradle.instrumentation.classloader.mapping.updateDaoMissingClasses
 
 class GeneratingMissingClassesClassLoader : ClassLoader(getSystemClassLoader()) {
@@ -12,7 +14,9 @@ class GeneratingMissingClassesClassLoader : ClassLoader(getSystemClassLoader()) 
             *deletionDaoMissingClasses,
             *insertionDaoMissingClasses,
             *updateDaoMissingClasses,
-            *selectDaoMissingClasses
+            *selectDaoMissingClasses,
+            *sqliteCopyOpenHelperMissingClasses,
+            *gmsMapping
         )
     }