+++ CHANGELOG.md
@@ -2,8 +2,21 @@
 
 ## Unreleased
 
+## 3.0.0-beta.3
+
+* Introduce the `includeProguardMapping` option to exclude the proguard logic, and deprecate `autoUpload` in favor of `autoUploadProguardMapping` (#240)
+* Feature: New File I/O auto-instrumentation (#249)
+* Feature: Add compile-time check for sentry-android SDK presence (#243)
+* Feature: New configuration option `tracingInstrumentation.features` to allow enabling/disabling certain features for auto-instrumentation (#245)
+
+## 3.0.0-beta.2
+
 * Fix: Correctly add the proguard UUID output directory to the source set (#226)
+* Feature: Make the ignoreXXX properties in SentryPluginExtension sets (#225)
 * Expose SentryPluginExtension.tracingInstrumentation (#229)
+* Ref: Change Room queries description to Dao class name (#232)
+* Fix: Log broken bytecode when build fails (#233)
+* Ref: Change db operation names (#237)
 
 ## 3.0.0-beta.1